eval/requirements.txt
openai>=0.27.0
transformers>=4.30.0
torch>=1.12.0
Pillow>=9.0.0
sacrebleu>=2.3.1
nltk>=3.6.0
tqdm>=4.60.0
numpy>=1.23.0
